Add files via upload
Run after hardening. This gathers md5sums of all files in the specified directories. Run this with a tarball of those directories. Run this again with a different save file name and when your services go down you can run a diff on the traverse filenames and replace the modified file with the one from the tarball.